It is a health chart manage program
Let assume that you are a gym instructor and you have 3 trainees, you have to maintain their food(diet) and exercises for each of them and update that time to time. With this program, there are 6 charts ( .txt files with their name ) for 3 people there is food and exercise charts. To run this program we need to create that also.
Three peoples are Harry, Hammad, Rohan.

If you want to learn python then visit #CodeWithHarry youtube channel. I also learn python from this channel.
This is a given task by #code with harry